refactor: split LogoutAlert scene's protocols into separated folder and files
