+++ Pharus/Scenes/Alerts/Logout Alert/LogoutAlertCoordinator.swift
@@ -7,11 +7,6 @@
 
 import UIKit
 
-protocol LogoutAlertFlow {
-    func closeModal()
-    func logout()
-}
-
 class LogoutAlertCoordinator: Coordinator {
 
     // MARK: - Properties
