+++ Pharus/Scenes/Alerts/Logout Alert/Protocols/LogoutAlertFlow.swift
@@ -1,0 +1,13 @@
+//
+//  LogoutAlertFlow.swift
+//  Pharus
+//
+//  Created by Victor Colen on 17/06/22.
+//
+
+import Foundation
+
+protocol LogoutAlertFlow {
+    func closeModal()
+    func logout()
+}
