+++ Pharus/Scenes/Alerts/Logout Alert/Protocols/LogoutAlertPresenterProtocol.swift
@@ -1,0 +1,13 @@
+//
+//  LogoutAlertPresenterProtocol.swift
+//  Pharus
+//
+//  Created by Victor Colen on 17/06/22.
+//
+
+import Foundation
+
+protocol LogoutAlertPresenterProtocol {
+    func closeModal()
+    func logout()
+}
