+++ Pharus/Scenes/Alerts/Logout Alert/Protocols/LogoutAlertViewDelegate.swift
@@ -1,0 +1,13 @@
+//
+//  LogoutAlertViewDelegate.swift
+//  Pharus
+//
+//  Created by Victor Colen on 17/06/22.
+//
+
+import Foundation
+
+protocol LogoutAlertViewDelegate: AnyObject {
+    func primaryButtonTapped()
+    func secondaryButtonTapped()
+}
